- Projects and About links in navbar
