+++ src/components/about/about.scss
@@ -1,0 +1,3 @@
+#about {
+    margin: 60px auto;
+}
